backend/requirements.txt
azure-core==1.37.0
azure-data-tables==12.7.0
azure-functions==1.24.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
isodate==0.7.2
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
yarl==1.22.0